Log all events and detail echo events
Previous calls to subscribe used `'/.*/'` (a string) instead of `/.*/`
(a regex) and failed to receive any events.

`EchoSubscriber` demonstrates a specific subscriber.  It uses specific
arguments to `#call`.  It only responds to `echo` events.

`LoggingSubscriber` demonstrates a more generic subscriber.  It uses
ActiveSupport's `Event` class to parse the event and offer value-added
information, such as `#duration`.  It responds to all events and shows
all that happens within Rails to answer an HTTP request.